+++ app/src/main/java/ru/bagrusss/revolutdemo/rates/RatesInteractor.kt
@@ -4,6 +4,7 @@ import io.reactivex.Completable
 import io.reactivex.Observable
 import ru.bagrusss.revolutdemo.mvvm.Interactor
 import ru.bagrusss.revolutdemo.rates.models.Rate
+import java.math.BigDecimal
 
 /**
  * Created by bagrusss on 13.08.2019
@@ -12,6 +13,6 @@ interface RatesInteractor: Interactor {
 
     val ratesChanges: Observable<List<Rate>>
 
-    fun rateChanged(rate: String, cost: Float): Completable
+    fun rateChanged(rate: String, cost: BigDecimal): Completable
 
 }
